+++ agent/src/main/java/com/teamscale/jacoco/agent/options/ArtifactoryConfig.java
@@ -2,8 +2,8 @@ package com.teamscale.jacoco.agent.options;
 
 import com.teamscale.client.CommitDescriptor;
 import com.teamscale.client.StringUtils;
-import com.teamscale.jacoco.agent.git_properties.GitPropertiesLocator;
-import com.teamscale.jacoco.agent.git_properties.InvalidGitPropertiesException;
+import com.teamscale.jacoco.agent.commit_resolution.git_properties.GitPropertiesLocator;
+import com.teamscale.jacoco.agent.commit_resolution.git_properties.InvalidGitPropertiesException;
 import okhttp3.HttpUrl;
 import org.conqat.lib.commons.collections.Pair;