+++ CHANGELOG.md
@@ -6,6 +6,7 @@ We use [semantic versioning](http://semver.org/):
 
 # Next Release
 - [feature] _tia-client_: add API to hash test data when creating ClusteredTestDetails
+- [feature] support for SAP NWDI application profiling
 
 # 20.0.0
 - [breaking change] This release requires Teamscale 5.9 or higher